+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - fixed a error in logs, when in the shipment volume calculation function none of the conditions are met
+- fixed errors being received when an error message is received instead of the terminals list
 
 ## [1.1.5] - 2024-03-20
 ### Fixed